Vector, Util-Klasse (für Hilfreiche Funktionen wie Abstand) und CollisionDetector hinzugefügt, PhysicalProperties um Radius erweitert (für einfache CollisionDetection)

+++ PhysicsEngine/src/de/engine/math/Point.java
@@ -2,6 +2,6 @@ package de.engine.math;
 
 public class Point 
 {
-	public float x;
-	public float y;
+	public double x;
+	public double y;
 }
